Feat: #67 - DF+Text 익스텐션 생성
