+++ 2024-MacC-M4-6princess/2024-MacC-M4-6princess/Source/App/_024_MacC_M4_6princessApp.swift
@@ -10,12 +10,6 @@ import SwiftUI
 @main
 struct _024_MacC_M4_6princessApp: App {
     @StateObject private var environmentModel = EnvironmentModel()
-//    init() {
-//        
-//        Font.registerFonts(fontName: "Hakgyoansim Geurimilgi OTF R")
-//        Font.registerFonts(fontName: "Pretendard-Bold")
-//        Font.registerFonts(fontName: "Pretendard-Regular")
-//    }
     let persistenceController = PersistenceController.shared
     
     var body: some Scene {
